+++ bin/make.js
@@ -9,8 +9,6 @@
 
 // Module dependencies.
 require('shelljs/make');
-var archiver=require('archiver');
-var fs=require('fs');
 var path=require('path');
 var pkg=require('../package.json');
 var util=require('util');
@@ -90,11 +88,6 @@ target.dist=function() {
     if(!test('-f', executable) && ('ProgramFiles(x86)' in env)) executable=path.join(env['ProgramFiles(x86)'], builder.binary);
     exec(util.format(builder.command, executable));
   });
-
-  var archive=archiver('zip');
-  archive.on('entry', function(entry) { echo('Pack:', entry.name); });
-  archive.pipe(fs.createWriteStream(config.output+'.zip'));
-  archive.bulk({ cwd: 'var/release', expand: true, src: '**/*' }).finalize();
 };
 
 /**