+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 This file contains highlights of what changes on each version of the [Crypt.cs](https://github.com/cedx/crypt.cs) application.
 
 #### Version 0.4.1
-- Updated the project license.
+- Changed the project license.
 
 #### Version 0.4.0
 - Initial release.