+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 This file contains highlights of what changes on each version of the [Crypt.cs](https://github.com/cedx/crypt.cs) application.
 
+#### Version 0.4.2
+- Using [Gulp.js](http://gulpjs.com) as build system.
+
 #### Version 0.4.1
 - Changed the project license.